src/backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
mysql-connector-python==8.1.0 